Added resume  as of July 2021
This resume is based on education and work experiences in health care, as a pharmacist.  I have worked in various pharmacy environments, and primarily involved in the COVID-19 vaccine task force as a contractor for the CDC.  I have not yet added my current position or additional skills that I have acquired since July, for HTML, CSS, Javascript, and git, but that will be coming soon.